+++ cvat/apps/annotation/transports/cvat/__init__.py
@@ -1,0 +1,1 @@
+from .importer import CVATImporter
